src/backend/scraper/facebook_group_scraper/requirements.txt
apify-client>=1.6.0
openai>=1.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
